# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PillMind API",
    "description": "PillMind 3.0 Kurumsal Entegrasyon, HL7 FHIR Uyumlu ve Güvenli İlaç Etkileşim Analiz API Sözleşmesi",
    "version": "3.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Lokal Geliştirme Sunucusu"
    }
  ],
  "paths": {
...
    "/api/explain": {
      "post": {
        "summary": "Yapay Zeka Destekli Klinik Açıklama Üretimi",
        "description": "İlaç etkileşimine dair hastaya yönelik sade Türkçe klinik açıklama ve hekim tavsiyelerini Gemini AI üzerinden üretir. Upstash Redis ile önbelleğe alınır.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "interactionId": { "type": "string", "example": "int-id-123" },
                  "drug1": { "type": "string", "example": "aspirin" },
                  "drug2": { "type": "string", "example": "warfarin" },
                  "severity": { "type": "string", "example": "high" },
                  "summary": { "type": "string", "example": "Artmış kanama riski." }
                },
                "required": ["interactionId", "drug1", "drug2", "severity", "summary"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Klinik açıklama nesnesi.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "girisCumlesi": { "type": "string" },
                    "klinikEtkiAciklamasi": { "type": "string" },
                    "hastalaraOneriler": { "type": "string" },
                    "hekimYonlendirmesi": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "description": "Eksik parametre" }
        }
      }
    },
...
  }
}